Benerin aset tetap & aset pengalihan

+++ sistemAset/app/Http/Controllers/AsetPengalihanController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use Illuminate\Support\Facades\DB;
 use Illuminate\Http\Request;
+use Session;
 
 class AsetPengalihanController extends Controller
 {
@@ -25,39 +26,76 @@ class AsetPengalihanController extends Controller
     public function store(Request $request)
     {
 
-        $nama_aset = DB::table('aset_tetaps')->where('id_Aset',$request->id_Aset)->first();
+        $aset_tetap = DB::table('aset_tetaps')->where('id_Aset',$request->id_Aset)->first();
 
-        DB::table('aset_pengalihan')->insert([
-            'nama_Aset' => $nama_aset->nama_Aset,
-            'jenis_Pengalihan' => $request->jenis_Pengalihan,
-            'jumlah' => $request->jumlah,
-            'lokasi_Pengalihan' => $request->lokasi_Pengalihan,
-            'id_Aset' => $request->id_Aset
-        ]);
+        if ($aset_tetap->jumlah > $request->jumlah) {
+            DB::table('aset_pengalihan')->insert([
+                'nama_Aset' => $aset_tetap->nama_Aset,
+                'jenis_Pengalihan' => $request->jenis_Pengalihan,
+                'jumlah' => $request->jumlah,
+                'lokasi_Pengalihan' => $request->lokasi_Pengalihan,
+                'id_Aset' => $request->id_Aset
+            ]);
 
-        DB::table('aset_tetaps')->where('id_Aset',$request->id_Aset)->delete();
-        
-        $aset_pengalihan = DB::table('aset_pengalihan')->count();
+            $jumlah_aset_tetap = $aset_tetap->jumlah - $request->jumlah;
+            DB::table('aset_tetaps')->where('id_Aset',$request->id_Aset)->update([
+                'jumlah' => $jumlah_aset_tetap
+            ]);
+    
+            // Hitung jumlah records aset pengalihan
+            $count_aset_pengalihan = DB::table('aset_pengalihan')->count();
+            DB::table('rekapitulasi')->where('id',6)->update([
+                'kuantitas' => $count_aset_pengalihan
+            ]);
+            
+            // Hitung jumlah records aset tetap
+            $count_aset_tetaps = DB::table('aset_tetaps')->count();
+            DB::table('rekapitulasi')->where('id',1)->update([
+                'kuantitas' => $count_aset_tetaps
+            ]);
 
-        DB::table('rekapitulasi')->where('id',6)->update([
-            'kuantitas' => $aset_pengalihan
-        ]);
+            return redirect('/AsetPengalihan');
+    
+        } else if ($aset_tetap->jumlah == $request->jumlah) {
+            DB::table('aset_pengalihan')->insert([
+                'nama_Aset' => $aset_tetap->nama_Aset,
+                'jenis_Pengalihan' => $request->jenis_Pengalihan,
+                'jumlah' => $request->jumlah,
+                'lokasi_Pengalihan' => $request->lokasi_Pengalihan,
+                'id_Aset' => $request->id_Aset
+            ]);
 
-        $count_aset_tetaps = DB::table('aset_tetaps')->count();
+            DB::table('aset_tetaps')->where('id_Aset',$request->id_Aset)->delete();
 
-        DB::table('rekapitulasi')->where('id',1)->update([
-            'kuantitas' => $count_aset_tetaps
-        ]);
+            $count_aset_pengalihan = DB::table('aset_pengalihan')->count();
+            DB::table('rekapitulasi')->where('id',6)->update([
+                'kuantitas' => $count_aset_pengalihan
+            ]);
+            
+            // Hitung jumlah records aset tetap
+            $count_aset_tetaps = DB::table('aset_tetaps')->count();
+            DB::table('rekapitulasi')->where('id',1)->update([
+                'kuantitas' => $count_aset_tetaps
+            ]);
 
-        return redirect('/AsetPengalihan');
+            // Hitung jumlah records aset tetap
+            $count_aset_tetaps = DB::table('aset_tetaps')->count();
+            DB::table('rekapitulasi')->where('id',1)->update([
+                'kuantitas' => $count_aset_tetaps
+            ]);
+
+            return redirect('/AsetPengalihan');
+        } else {
+            Session::flash('error','Stok tidak tersedia');
+		    return redirect('/AsetPengalihan/tambah');
+        }
     }
     
-    public function edit($id_Aset)
+    public function edit($id_Aset_Pengalihan)
     {
-        $aset_pengalihan = DB::table('aset_pengalihan')->where('id_Aset',$id_Aset)->get();
-        $aset_tetaps = DB::table('aset_tetaps')->where('id_Aset',$id_Aset)->get();
+        $aset_pengalihan = DB::table('aset_pengalihan')->where('id_Aset_Pengalihan',$id_Aset_Pengalihan)->get();
 
-        return view('asetPengalihan.editAsetPengalihan',['aset_pengalihan' => $aset_pengalihan, 'aset_tetaps' => $aset_tetaps]);
+        return view('asetPengalihan.editAsetPengalihan',['aset_pengalihan' => $aset_pengalihan]);
     }
 
     public function update(Request $request)
@@ -72,12 +110,23 @@ class AsetPengalihanController extends Controller
         return redirect('/AsetPengalihan');
     }
 
-    public function hapus($id_Aset)
+    public function hapus($id_Aset_Pengalihan)
     {
-        DB::table('aset_pengalihan')->where('id_Aset',$id_Aset)->delete();
+        $aset_pengalihan = DB::table('aset_pengalihan')->where('id_Aset_Pengalihan',$id_Aset_Pengalihan)->first();
+        $aset_tetaps = DB::table('aset_tetaps')->where('id_Aset',$aset_pengalihan->id_Aset)->first();
+
+        if ($aset_tetaps != null) {
+            $stok_aset_tetaps = $aset_pengalihan->jumlah + $aset_tetaps->jumlah;
+
+            DB::table('aset_tetaps')->where('id_Aset',$aset_tetaps->id_Aset)->update([
+                'jumlah' => $stok_aset_tetaps
+            ]);
+        }
+
+        DB::table('aset_pengalihan')->where('id_Aset_Pengalihan',$id_Aset_Pengalihan)->delete();
 
         $count_aset_pengalihan = DB::table('aset_pengalihan')->count();
-
+        
         DB::table('rekapitulasi')->where('id',6)->update([
             'kuantitas' => $count_aset_pengalihan
         ]);
